FMTX v1
This version clears the v unvisited in update function. Also uses an if condition to not put
nodes that has inf cost to v open (because you cleared the v unvisted we don't have access to it).
The reason for clearing is to not handle that in the else if part of the plan function or else we have to put the parent node of the promising node (the one that got into v unvisted in the else if part) into v open.
so all in all there is not direct handling of those promising nodes so we just clear them! which i guess its not good for the later early exit implementation.