FMTX v2
This version doesnt clear the v unvisited node in update instead in the else if part we put the parent of the promising node
into the v open so that eventually it gets the same result and get deleted from v unvisted in the main connection procedure.
note that we dont invalidate anything in the else if part because there is no need since if it gets any better the change will happen in the main if condition after obstalce check.
This version opens up new improvment for early exit but still a little bit cluttered.