Add early exit
This version got really complicated I put back the v unvisted earse in the main if but i still
need to test see if the previous idea holds or not
also the main problem that got me to make the code a little bit complicated is that we need to make sure
that no node that has inf value get into v open! the best way is not to just put and if condition and ignore them. the best way is to track all the inf nodes, like in handleAddObstalce function
orphan nodes loop needs to have v_open_set.erase(orphan) line. these are because of a design choice of trying to hold on to the v open heap and set during the early exit.
the next design test would be to erase them all because handling them is quite complicated!
Also maybe its time to use the priority queue with an update/remove feature. but i try to avoid that as much as possible.