Remove samples in obstalce from v unvisited nodes earlier than the v open push loop
Well choosing the union of added and removed nodes as the starting point of the expansion was not enough.
Imgine the two sides of a plier and in handleAddObstalce when some nodes become orphans they need to get connected to the other side (the branch on the opposite side), and choosing vopen nodes to only start from the vicinity of the obstalce change is not gonna put those node at the edge of the tree
in the vOpen, hence, we get sub optimal solutions. so we need to loop through all the v unvisted nodes and put all their neighbors into v open because we never know which side of the tree is gonna help us and its not necessarily the current direction of the epxnasion that gonna helps us (although this is not true for remove obstacle and its only true for add obstalce).
so since we are looping through the v unvisted nodes every time I thought to erase the samples on obstalces from v unvisted before this loop so that we wouldn't waste so much time here.

Another piece of info i want to give is this:
one might ask why do you put orphan nodes into v_unvisited_set when you have a mechanism in the main loop to find these automatically?!
The reason is these help the finding of the v open nodes later in the update obstalce sample function
If we only rely on that mechansim we can't find connections to other branches because we are blind to see other branches! like on the other side of the tree
Imagine the one side of the plier and some nodes get better cost if they get connected to the other tip of the plier but since we didn't put the other side nodes into v open we never know!

So that condtion in the main loop is just for one direction expansion and is good for the nodes that gor removed from the obstalce--> Although its a reasonable question here also to ask ourselves why its not the case
for the remove obstlace to now know their v open at first!
the difference between addObstalce and removeObstalce is adding and obstalce most certainly adds cost to orphan nodes
but removing an obstlace most certainly reduces cost of the neighbor nodes! and reducing happens in the current branch and direction of the expansion that happens in dijkstra like (like fmtx) algorithm
so we don't need to worry about the other side of plier (per say!) because they are gonna connect to us! not us connecting to them (and by "us" i mean the current direction of the expansion)