Test invalid_neighbor_nodes feature performance
using nested hash is computationally expensive
using bit mask of bool requires to have num_samples*num_samples*1byte. also i need resizing
using time stamp approach to not resize is good but needs vector of vector of intes instead of the above approach which is vector of vector of bool. so it needs num_samples_*num_samples*4bytes which gets huge for large graphs!
for the next commit i will test vector of unordered map to have a balance in computaion and memory usage