Clean up some lines and a good comment!
I don't think we need to track the max_edge in fmtx because we know beforehand
what could be the upper bound for the max_length and its neighborhood radius!
but i kept the implementation for now but i commented it!

Also i added the feature to be able to add new goals from rviz interactively