Use flat map for rrtx
Since the lookup is too much and i want to test small number of nodes im gonna
use this even though insertion time is O(log n) but since the array is sorted and we do lots of iteration over neighbors its more cpu friendly
the experiment also shows its faster as opposed to unordered map. at least for 5000 nodes it was.