Add collision check to dynamic procedure as the second approach to ignore_sample
Also fixed a bug on gazebo obstacle checker to save an snapshot of the obstacle
position since it continually got updated as the plan while loop was working

prune true means do an obstacle check in the dynamic procedure and not obstacle check in plan
prune false means use the delayed obstacle approach of fmt but its slow since we are checking all the edges!
but in prune false we check all the local edges (node to all its neighbor) and invalidate them setting their dist to infinity
but maybe the delayed approach works better in high dimension but maybe not!