Fix the percision problem that will happen when you update the samples_in_obstalces to current before the for loop
This problem is happening because in remove obstalce procedure you put some nodes that already have parent
in the in_unvisited and then in plan() procedure it reaches the core cost update and the newCost is the same as the old cost and it doesnt update and keeps its last state
which is the same state if it even updates but the thing is that in_unvisited doesnt get to be false so in the next iteration in the handle add or remove obstalce the heap addition will experience some problems because it ignores nodes in v unvisted!
so i had to put some parent check in the remove obstalce so that if the node already have a parent then ignore that
to visualize the problem you can comment the parent check condtion and put one obstalce in env
then move it far away and you can see some of the surrounding node persist to be in_unvisted (ofcourse you have to use visualizeHeapAndUnvisted function in the update sample obstalce function)

In the next commit im gonna completely erase the usage of in_unvisited. there is no need for it and it served its purpose!