Test the unvisited condtion vs cost condtion based on number of obstacle checks
Just comment the 1st/2nd condtion and test (ofcourse you should test with the presence of obstalce)
This is a test in static envrionment to see if my condition would cause more obstacle test or not
It didn't because those condtions are somehow equivalent because the whole idea of using unvisited condtion is to go through each node once and
by using the cost condition after a node has updated then it never satisfies the (x->getCost() > (z->getCost() + cost_to_neighbor.distance ) condtion again in the same for loop
and even in the next while loop iteration. its obvious why its not in the same for loop because we are using the same z node. and its also obvious why its not gonna cause issue in the next while iteration because the next z we pop is surely greater than the previous z node so (x->getCost() > (z->getCost() + cost_to_neighbor.distance ) only gets called until x is connected and then it will never allow us to move passed that condtion for that x node again.