Back to cost condtion
About my last commit's comment:
One might ask you are connecting the x not to z but to the best neighbor(considering obstacles)
so if we get passed this (x->getCost() > (z->getCost() + cost_to_neighbor.distance ) ) until we connect x
the next z might trigger the condtion to get passed this. IT MIGHT BE TRUE and i need to test but in the worst case scenrio i only have to
calculated best_neighbor_cost and check if we again get the same cost and can early exit before obstalce check. BUT I NEED TO TEST