Add KNN feature to fmtx
fmtx can use this to reduce its rewire searches that happen in plan function significantly
something rrtx cant use persumably due its not unless we forget about the cullNeighbor idea maybe
and that may have require us to change some part of the algorithm it self.
maybe it can be integrated to rrtx if i tried. CullNeighbor needs to be changed right of the bat.