Add different modes to RRTX
mode 1 is node centric and stores current and previous conflicting nodes. kd tree query is less but more obstalce checks
mode 2 is fully obstacle centric and loop over current and previous obstacles. kd tree query is doubled but obstacle checks are reduced and specific
mode 3 is node centric with a map that associates conflicting nodes to their corresponding obstacles so that we could have reduced obstacle checks and keep the less kd tree query of mode 1

mode 2 is the default approach of rrtx. mode 1 is the one that i initially designed with and mode 3 is the optimization of mode 1.
I observed mode 2 is faster than mode 1. mode 3 is also faster than mode 1.
In geometric case i observed with 10,20 dynamic obstalces, mode 2 is faster than mode 3, but with 30 obstacles mode 3 is faster!.
caveat with mode 3 is that it needs to handle a hash map structure. if i could make it faster then i think mode 3 is better but for comparison purposes
I use mode 2.