Add obstacle centric approach to FMTx
Mind that obstalce centric approach is not the default approach for fmtx because it doesnt use the delayed collision checking
When obstalce check is very cheap. this approach is okayish!
All im trying to do is to add all these features like knn (to reduce number of times we move throguh the inner for loop of the plan function) and prune mode (mode 2)and say
FMTx is flexible to use all this stuff!